correction du type pris dans la multiplication par un double et non un entier

See merge request romeo_jacquemin/ray-tracing!9

+++ src/main/fr/univartois/iutl/info/raytracing/numeric/Color.java
@@ -49,7 +49,7 @@ public class Color {
 	 * @return Color
 	 * multiplicate the rgb coordinates with an integer
 	 */
-	public Color multiplication(int d) {
+	public Color multiplication(double d) {
 		return new Color(triplets.multiplication(d));
 	}
